refactor(hapi): model a local call's outcome as the two shapes it has
`run_local` took `&self` and never read it; it is a free function. `Local`
held two `Option`s, so a fourth state — refused by nothing, carrying nothing —
was representable and meant nothing. It is now `Answered` or `Refused`, the
same split `queries::answered` and `queries::refused` already make on the
header.

`Curve` is used by one function and now lives in it. `HBAR_EQUIVALENT` and
`CENT_EQUIVALENT` were `i32` because protobuf wants `i32`; they are counts,
so they are `u32` and the cast happens at the protobuf boundary.

+++ src/hapi/render.rs
@@ -11,8 +11,8 @@ use crate::state::{
 fn exchange_rate(chain: &Chain) -> proto::ExchangeRateSet {
     let expires = chain.latest_block().consensus_timestamp.secs + EXCHANGE_RATE_VALID_SECS;
     let rate = proto::ExchangeRate {
-        hbar_equiv: HBAR_EQUIVALENT,
-        cent_equiv: CENT_EQUIVALENT,
+        hbar_equiv: HBAR_EQUIVALENT as i32,
+        cent_equiv: CENT_EQUIVALENT as i32,
         expiration_time: Some(proto::TimestampSeconds {
             seconds: expires as i64,
         }),
